+++ libavutil/x86/crc.h
@@ -28,10 +28,11 @@
 #include "libavutil/cpu.h"
 #include "libavutil/crc.h"
 #include "libavutil/intreadwrite.h"
-#include "libavutil/reverse.h"
 #include "libavutil/x86/cpu.h"
 
 #if HAVE_CLMUL_EXTERNAL
+#include "libavutil/crc_internal.h"
+
 FF_VISIBILITY_PUSH_HIDDEN
 uint32_t ff_crc_clmul(const AVCRC *ctx, uint32_t crc,
                       const uint8_t *buffer, size_t length);
@@ -103,47 +104,6 @@ static const AVCRC crc_table_clmul[AV_CRC_MAX][17] = {
         0xcfffbffe, 0x1, 0x14003, 0x0,
     },
 };
-
-static uint64_t reverse(uint64_t p, unsigned int deg)
-{
-    uint64_t ret = 0;
-    int i;
-    for (i = 0; i < (deg / 8); i += 1) {
-        ret = (ret << 8) | (ff_reverse[p & 0xff]);
-        p >>= 8;
-    }
-    int rem = (deg + 1) - 8 * i;
-    ret = (ret << rem) | (ff_reverse[p & 0xff] >> (8 - rem));
-    return ret;
-}
-
-static uint64_t xnmodp(unsigned n, uint64_t poly, unsigned deg, uint64_t *div, int bitreverse)
-{
-    uint64_t mod, mask, high;
-
-    if (n < deg) {
-        *div = 0;
-        return poly;
-    }
-    mask = ((uint64_t)1 << deg) - 1;
-    poly &= mask;
-    mod = poly;
-    *div = 1;
-    deg--;
-    while (--n > deg) {
-        high = (mod >> deg) & 1;
-        *div = (*div << 1) | high;
-        mod <<= 1;
-        if (high)
-            mod ^= poly;
-    }
-    uint64_t ret = mod & mask;
-    if (bitreverse) {
-        *div = reverse(*div, deg) << 1;
-        return reverse(ret, deg) << 1;
-    }
-    return ret;
-}
 
 static inline void crc_init_x86(AVCRC *ctx, int le, int bits, uint32_t poly, int ctx_size)
 {
